requirements.txt
translators
azure-ai-translation-text
python-dotenv